Add reward entity to schema, deprecate mission and action entities
