+++ schema.graphql
@@ -147,6 +147,26 @@ type User @entity {
   updatedAtBlock: BigInt!
 }
 
+type Reward @entity (immutable: true) {
+  id: ID!
+  app: App!
+  rewardId: String!
+  rewardType: String!
+  metadataURI: String!
+  user: User!
+  token: FungibleToken
+  tokenAmount: BigInt!
+  badge: Badge
+  badges: [BadgeToken!]
+  badgeCount: Int!
+  transactionHash: String!
+  createdAt: BigInt!
+  createdAtBlock: BigInt!
+}
+
+"""
+DEPRECATED: The `Action` entity has been deprecated as please query the `Reward` entity with `rewardType: "ACTION"` instead.
+"""
 type Action @entity {
   id: ID!
   metadata: ActionMetadata
@@ -157,12 +177,18 @@ type Action @entity {
   createdAtBlock: BigInt!
 }
 
+"""
+DEPRECATED: The `ActionMetadata` entity has been deprecated as please query the `Reward` with `rewardType: "ACTION"` instead.
+"""
 type ActionMetadata @entity {
   id: ID!
   name: String
   URI: String!
 }
 
+"""
+DEPRECATED: The `MissionFungibleToken` entity has been deprecated as please query the `Reward` entity instead.
+"""
 type MissionFungibleToken @entity {
   id: ID!
   token: FungibleToken!
@@ -170,12 +196,18 @@ type MissionFungibleToken @entity {
   amount_rewarded: BigInt!
 }
 
+"""
+DEPRECATED: The `MissionMetadata` entity has been deprecated as please query the `Reward` entity with `rewardType: "MISSION"` instead.
+"""
 type MissionMetadata @entity {
   id: ID!
   name: String
   URI: String!
 }
 
+"""
+DEPRECATED: The `Mission` entity has been deprecated as please query the `Reward` entity with `rewardType: "MISSION"` instead.
+"""
 type Mission @entity {
   id: ID!
   metadata: MissionMetadata
